examples/requirements.txt
requests
pythorhead
pythonseer>=0.1.3
python-dotenv